+++ src/str_utils.c
@@ -22,7 +22,7 @@ char* trimNewline(char* str) {
 
 
 // Function to split a line into tokens by a delimiter
-char** split(const char* line, const char* delimiter, int* count) {
+char** split(const char* line, const char* delimiter, size_t* count) {
     char* pos = strdup(line);
     char* token = NULL;
     char** tokens = NULL;
